Refact archive methods
